chore: rename package `mashup.mmz` to `com.piikii`
