+++ piikii-application/src/main/kotlin/com/piikii/application/domain/model/User.kt
@@ -1,0 +1,6 @@
+package com.piikii.application.domain.model
+
+class User(
+    val loginId: String
+) {
+}
