+++ piikii-application/src/main/kotlin/com/piikii/application/port/input/UserUsecase.kt
@@ -1,0 +1,7 @@
+package com.piikii.application.port.input
+
+import com.piikii.application.domain.model.User
+
+interface UserUsecase {
+    fun save(loginId: String): User
+}
