+++ piikii-application/src/main/kotlin/com/piikii/application/port/output/eventbroker/Event.kt
@@ -1,0 +1,6 @@
+package com.piikii.application.port.output.eventbroker
+
+interface Event {
+
+    fun getTopic(): Event
+}
