+++ piikii-application/src/main/kotlin/com/piikii/application/port/output/eventbroker/UserConsumer.kt
@@ -1,0 +1,5 @@
+package com.piikii.application.port.output.eventbroker
+
+interface UserConsumer {
+    fun execute(): Event
+}
