+++ piikii-application/src/main/kotlin/com/piikii/application/port/output/eventbroker/UserProducer.kt
@@ -1,0 +1,5 @@
+package com.piikii.application.port.output.eventbroker
+
+interface UserProducer {
+    fun execute(event: Event)
+}
