+++ piikii-application/src/main/kotlin/com/piikii/application/port/output/eventbroker/UserUpdatedEvent.kt
@@ -1,0 +1,8 @@
+package com.piikii.application.port.output.eventbroker
+
+class UserUpdatedEvent : Event {
+
+    override fun getTopic(): Event {
+        TODO("Not yet implemented")
+    }
+}
