--- piikii-application/src/main/kotlin/mashup/mmz/application/domain/model/User.kt
@@ -1,6 +1,0 @@
-package mashup.mmz.application.domain.model
-
-class User(
-    val loginId: String
-) {
-}
