--- piikii-application/src/main/kotlin/mashup/mmz/application/port/input/UserUsecase.kt
@@ -1,7 +1,0 @@
-package mashup.mmz.application.port.input
-
-import mashup.mmz.application.domain.model.User
-
-interface UserUsecase {
-    fun save(loginId: String): User
-}
