--- piikii-application/src/main/kotlin/mashup/mmz/application/port/output/eventbroker/Event.kt
@@ -1,6 +1,0 @@
-package mashup.mmz.application.port.output.eventbroker
-
-interface Event {
-
-    fun getTopic(): Event
-}
