--- piikii-application/src/main/kotlin/mashup/mmz/application/port/output/eventbroker/UserConsumer.kt
@@ -1,5 +1,0 @@
-package mashup.mmz.application.port.output.eventbroker
-
-interface UserConsumer {
-    fun execute(): Event
-}
