--- piikii-application/src/main/kotlin/mashup/mmz/application/port/output/eventbroker/UserCreatedEvent.kt
@@ -1,8 +1,0 @@
-package mashup.mmz.application.port.output.eventbroker
-
-class UserCreatedEvent : Event {
-
-    override fun getTopic(): Event {
-        TODO("Not yet implemented")
-    }
-}
