--- piikii-application/src/main/kotlin/mashup/mmz/application/port/output/eventbroker/UserProducer.kt
@@ -1,5 +1,0 @@
-package mashup.mmz.application.port.output.eventbroker
-
-interface UserProducer {
-    fun execute(event: Event)
-}
