--- piikii-application/src/main/kotlin/mashup/mmz/application/port/output/eventbroker/UserUpdatedEvent.kt
@@ -1,8 +1,0 @@
-package mashup.mmz.application.port.output.eventbroker
-
-class UserUpdatedEvent : Event {
-
-    override fun getTopic(): Event {
-        TODO("Not yet implemented")
-    }
-}
