+++ piikii-bootstrap/src/main/kotlin/com/piikii/bootstrap/BootstrapApplication.kt
@@ -1,0 +1,12 @@
+package com.piikii.bootstrap
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication(scanBasePackages = ["com.piikii"])
+class BootstrapApplication
+
+fun main(args: Array<String>) {
+     System.setProperty("spring.config.location", "classpath:/database-config/,classpath:/cache-config/,classpath:/broker-config/");
+    runApplication<BootstrapApplication>(*args)
+}
