+++ piikii-bootstrap/src/test/kotlin/com/piikii/bootstrap/BootstrapApplicationTests.kt
@@ -1,0 +1,13 @@
+package com.piikii.bootstrap
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class BootstrapApplicationTests {
+
+    @Test
+    fun contextLoads() {
+    }
+
+}
