--- piikii-bootstrap/src/test/kotlin/mashup/mmz/bootstrap/BootstrapApplicationTests.kt
@@ -1,13 +1,0 @@
-package mashup.mmz.bootstrap
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class BootstrapApplicationTests {
-
-    @Test
-    fun contextLoads() {
-    }
-
-}
