+++ piikii-common/exception/src/main/kotlin/com/piikii/common/exception/BaseException.kt
@@ -1,0 +1,4 @@
+package com.piikii.common.exception
+
+class BaseException {
+}
