--- piikii-common/exception/src/main/kotlin/mashup/mmz/common/exception/ApiBizException.kt
@@ -1,4 +1,0 @@
-package mashup.mmz.common.exception
-
-class ApiBizException {
-}
