+++ piikii-output-cache/caffeine/src/main/kotlin/Main.kt
@@ -1,4 +1,4 @@
-package mashup.mmz
+package com.piikii
 
 fun main() {
     println("Hello World!")
