+++ piikii-output-eventbroker/kafka/src/main/kotlin/com/piikii/output/eventbroker/kafka/KafkaConfig.kt
@@ -1,0 +1,7 @@
+package com.piikii.output.eventbroker.kafka
+
+import org.springframework.kafka.annotation.EnableKafka
+
+@EnableKafka
+class KafkaConfig {
+}
