+++ piikii-output-eventbroker/kafka/src/main/kotlin/com/piikii/output/eventbroker/kafka/user/UserKafkaProducer.kt
@@ -1,0 +1,13 @@
+package com.piikii.output.eventbroker.kafka.user
+
+import com.piikii.application.port.output.eventbroker.Event
+import com.piikii.application.port.output.eventbroker.UserProducer
+import org.springframework.stereotype.Component
+
+@Component
+class UserKafkaProducer : UserProducer {
+
+    override fun execute(event: Event) {
+        TODO("Not yet implemented")
+    }
+}
