--- piikii-output-eventbroker/kafka/src/main/kotlin/mashup/mmz/output/eventbroker/kafka/KafkaConfig.kt
@@ -1,7 +1,0 @@
-package mashup.mmz.output.eventbroker.kafka
-
-import org.springframework.kafka.annotation.EnableKafka
-
-@EnableKafka
-class KafkaConfig {
-}
