+++ piikii-output-persistence/postgresql/src/main/kotlin/com/piikii/output/persistence/postgresql/PostgreSQLConfig.kt
@@ -1,0 +1,13 @@
+package com.piikii.output.persistence.postgresql
+
+import org.springframework.boot.autoconfigure.domain.EntityScan
+import org.springframework.context.annotation.Configuration
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories
+
+@Configuration
+@EnableJpaAuditing
+@EntityScan(basePackages = ["mashup.mmz"])
+@EnableJpaRepositories(basePackages = ["com.piikii"])
+class PostgreSQLConfig {
+}
