+++ piikii-output-persistence/postgresql/src/main/kotlin/com/piikii/output/persistence/postgresql/user/UserJpaRepository.kt
@@ -1,0 +1,6 @@
+package com.piikii.output.persistence.postgresql.user
+
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface UserJpaRepository : JpaRepository<UserEntity, Long> {
+}
