--- piikii-output-persistence/postgresql/src/main/kotlin/mashup/mmz/output/persistence/postgresql/BaseEntity.kt
@@ -1,23 +1,0 @@
-package mashup.mmz.output.persistence.postgresql
-
-import jakarta.persistence.*
-import org.springframework.data.annotation.CreatedDate
-import org.springframework.data.annotation.LastModifiedDate
-import org.springframework.data.jpa.domain.support.AuditingEntityListener
-import java.time.LocalDateTime
-
-@EntityListeners(AuditingEntityListener::class)
-@MappedSuperclass
-abstract class BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val id: Long = 0L
-
-    @CreatedDate
-    @Column(name = "created_at", nullable = false, updatable = false)
-    var createdAt: LocalDateTime = LocalDateTime.now()
-
-    @LastModifiedDate
-    @Column(name = "modified_at", nullable = false)
-    var modifiedAt: LocalDateTime = LocalDateTime.now()
-}
