--- piikii-output-persistence/postgresql/src/main/kotlin/mashup/mmz/output/persistence/postgresql/PostgreSQLConfig.kt
@@ -1,13 +1,0 @@
-package mashup.mmz.output.persistence.postgresql
-
-import org.springframework.boot.autoconfigure.domain.EntityScan
-import org.springframework.context.annotation.Configuration
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories
-
-@Configuration
-@EnableJpaAuditing
-@EntityScan(basePackages = ["mashup.mmz"])
-@EnableJpaRepositories(basePackages = ["mashup.mmz"])
-class PostgreSQLConfig {
-}
