--- piikii-output-persistence/postgresql/src/main/kotlin/mashup/mmz/output/persistence/postgresql/user/UserJpaRepository.kt
@@ -1,6 +1,0 @@
-package mashup.mmz.output.persistence.postgresql.user
-
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface UserJpaRepository : JpaRepository<UserEntity, Long> {
-}
