+++ settings.gradle.kts
@@ -1,7 +1,7 @@
 plugins {
     id("org.gradle.toolchains.foojay-resolver-convention") version "0.5.0"
 }
-rootProject.name = "mmz"
+rootProject.name = "piikii"
 
 include("piikii-application")
